Optimized function
Finally fixed and added an optimized function. For big arrays with step 1 it is 5 times faster even though it looks very complicated. Currently there is no way to reset static variables for using with a different array, but if there would be such a need, I can add that. Also fixed a bug in the unoptimized version.